+++ src/convert/from_cosmx_to_spatialexperiment/script.R
@@ -2,7 +2,7 @@ library(SpatialExperimentIO)
 
 ### VIASH START
 par <- list(
-  input = "resources_test/cosmx/Lung5_Rep2_tiny",
+  input = "resources_test/cosmx/test2.zip",
   add_tx_path = TRUE,
   add_polygon_path = FALSE,
   add_fov_positions = TRUE,
@@ -11,16 +11,41 @@ par <- list(
   ),
   output = "spe_cosmx_test.rds"
 )
+meta <- list(
+  resources_dir = "src/utils/"
+)
 ### VIASH END
 
+source(paste0(meta$resources_dir, "/unzip_archived_folder.R"))
+
+cat("Reading input data...")
+if (tools::file_ext(par$input) == "zip") {
+  expected_file_patterns <- c(
+    "*.csv",
+    "*.parquet"
+  )
+  tmp_dir <- extract_selected_files(
+    par$input,
+    members = expected_file_patterns
+  )
+  cosmx_output_bundle <- file.path(
+    tmp_dir,
+    tools::file_path_sans_ext(basename(par$input))
+  )
+} else {
+  cosmx_output_bundle <- par$input
+}
+
+cat("Setting parameters...")
 if (par$add_polygon_path == FALSE && par$add_tx_path == FALSE) {
   add_parquet_paths <- FALSE
 } else {
   add_parquet_paths <- TRUE
 }
 
+cat("Converting to SpatialExperiment...")
 spe <- readCosmxSXE(
-  dirName = par$input,
+  dirName = cosmx_output_bundle,
   returnType = "SPE",
   countMatPattern = "exprMat_file.csv",
   metaDataPattern = "metadata_file.csv",
@@ -33,4 +58,5 @@ spe <- readCosmxSXE(
   altExps = par$alternative_experiment_features
 )
 
+cat("Saving output...")
 saveRDS(spe, file = par$output)