+++ src/convert/from_xenium_to_spatialdata/script.py
@@ -1,9 +1,11 @@
 import sys
 from spatialdata_io import xenium
+import zipfile
+from pathlib import Path
 
 ## VIASH START
 par = {
-    "input": "./resources_test/xenium_tiny",
+    "input": "xenium_tiny.zip",
     "output": "./test/xenium_tiny.zarr",
     "cells_boundaries": True,
     "nucleus_boundaries": True,
@@ -22,12 +24,29 @@ meta = {"resources_dir": "src/utils"}
 
 sys.path.append(meta["resources_dir"])
 from setup_logger import setup_logger
+from unzip_archived_folder import unzip_archived_folder
 
 logger = setup_logger()
 
 logger.info("Reading in Xenium data...")
+
+if zipfile.is_zipfile(par["input"]):
+    required_file_patterns = [
+        "**/experiment.xenium",
+        "**/nucleus_boundaries.parquet",
+        "**/cell_boundaries.parquet",
+        "**/transcripts.parquet",
+        "**/cell_feature_matrix.h5",
+        "**/cells.parquet",
+        "**/morphology_mip.ome.tif",
+        "**/morphology_focus.ome.tif",
+    ]
+    xenium_output_bundle = unzip_archived_folder(par["input"])
+else:
+    xenium_output_bundle = Path(par["input"])
+
 sdata = xenium(
-    par["input"],
+    xenium_output_bundle,
     cells_boundaries=par["cells_boundaries"],
     nucleus_boundaries=par["nucleus_boundaries"],
     cells_as_circles=par["cells_as_circles"],